await scripts by chunk order
